Modified for binding to all available interfaces (0.0.0.0) and easier instructions

+++ 6.3.0/build.sh
@@ -1,1 +1,10 @@
+#!/bin/bash
+
+if [ ! -e files/jboss-eap-6.3.0.zip ]
+then
+   echo "You should get required EAP binary before trying to build this image."
+   echo "Binary is available at: https://access.redhat.com/jbossnetwork/restricted/softwareDownload.html?softwareId=32483&product=appplatform"
+   exit 255
+fi
+
 docker build --rm -t jboss/eap:6.3.0  .
